requirements.txt
asarPy==1.0.1
asttokens==3.0.0
bcrypt==4.3.0
blinker==1.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
decorator==5.2.1
executing==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
greenlet==3.2.1
idna==3.10
ijson==3.4.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib-inline==0.1.7
numpy==2.2.5
opencv-python==4.11.0.86
packaging==25.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
prompt_toolkit==3.0.51
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pytesseract==0.3.13
requests==2.32.3
SQLAlchemy==2.0.40
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.13.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
